Refactor #57 로그인 수정 - 로그인 매니저 수정 - 프록시 수정 - 핸들러 분리
